backend/requirements.txt
alembic==1.13.1
annotated-types==0.7.0
anyio==4.8.0
bcrypt==4.0.1
cachetools==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
coverage==7.5.3
cryptography==44.0.2
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
factory-boy==3.3.0
Faker==24.4.0
fastapi==0.115.8
fastapi-cli==0.0.7
gevent==24.11.1
gevent-eventemitter==2.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
Jinja2==3.1.5
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
passlib==1.7.4
protobuf==3.20.3
psycopg2-binary==2.9.10
pyasn1==0.4.8
pycryptodomex==3.21.0
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pytest==8.1.1
pytest-asyncio==0.23.5
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-xdist==3.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.4.0
python-multipart==0.0.20
PyJWT==2.8.0
PyYAML==6.0.2
requests==2.32.3
requests-mock==1.11.0
rich==13.9.4
rich-toolkit==0.13.2
rsa==4.9
setuptools==75.8.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.30
starlette==0.45.3
steam==1.4.4
testcontainers==3.7.1
typer==0.15.1
typing_extensions==4.12.2
typing-inspect==0.4.0
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
vdf==3.4
watchfiles==1.0.4
websockets==14.2
wheel==0.45.1
zope.event==5.0
zope.interface==7.2
prometheus-client==0.20.0
psutil==5.9.8
opentelemetry-api==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-instrumentation-fastapi==0.46b0
opentelemetry-exporter-prometheus==0.46b0
